Fuzzy matching for team names
Nobody's actually going to type team names exactly how they appear in
the API; this change means that, if it doesn't get an exact match, it'll
intelligently guess.

First it will look for any team names that have an edit distance of less
than 5 from the given name – so that means 5 insertions, deletions, or
substitutions. That will catch simple typos, the omission of "FC" at the
end of a name, or something like that.

Then it will use White similarity, which looks at adjacent letter pairs
common to both strings. It will pick the most similar match of any that
score over 0.5.

If it still doesn't have a match that point, it will give up.

Further reading on White Similarity:
http://www.catalysoft.com/articles/StrikeAMatch.html